// BOTAIML.VisionBot.Monitoring.Web/ClientApp/src/app/api/models/application-permissions.ts
/* tslint:disable */
/* eslint-disable */
export enum ApplicationPermissions {
  UserRead = 'UserRead',
  UserCreate = 'UserCreate',
  UserUpdate = 'UserUpdate',
  RoleRead = 'RoleRead',
  RoleCreate = 'RoleCreate',
  RoleUpdate = 'RoleUpdate',
  DoorAccess = 'DoorAccess',
  SubCenterPowerAccess = 'SubCenterPowerAccess',
  AreaInsideAccess = 'AreaInsideAccess',
  AreaOutsideAccess = 'AreaOutsideAccess',
  GensetAreaAccess = 'GensetAreaAccess',
  ManageLiveStream = 'ManageLiveStream',
  ReadLiveStream = 'ReadLiveStream',
  ViewLiveStream = 'ViewLiveStream',
  EnrolmentRead = 'EnrolmentRead',
  EnrolmentCreate = 'EnrolmentCreate',
  EnrolmentDelete = 'EnrolmentDelete',
  EnrolmentUpdate = 'EnrolmentUpdate',
  AlertsRead = 'AlertsRead',
  ReportRead = 'ReportRead',
  ReportCreate = 'ReportCreate',
  ReportUpdate = 'ReportUpdate',
  ReportView = 'ReportView',
  ReportRun = 'ReportRun'
}